requirements.txt
appnope==0.1.2django-q==1.3.4
Babel==2.9.0
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
decorator==4.4.2
Django==3.1.4
django-extensions==3.1.0
django-filter==2.4.0
djangorestframework==3.12.2
django-q==1.3.4
flake8==3.8.4
importlib-metadata==3.3.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
mccabe==0.6.1
numpy==1.19.5
pandas==1.1.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.3
python-dateutil==2.8.1
pytz==2020.5
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
telebot==0.0.4
Telethon==1.19.1
